开启2025.12

+++ taotao-cloud-message-application/src/main/java/com/taotao/cloud/message/application/shared/monitor/CpuInfo.java
@@ -16,11 +16,11 @@
 
 package com.taotao.cloud.message.application.shared.monitor;
 
+import cn.hutool.core.util.NumberUtil;
 import java.math.BigDecimal;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
-import org.dromara.hutool.core.math.NumberUtil;
 
 /**
  * CPU相关信息
@@ -65,15 +65,15 @@ public class CpuInfo {
     }
 
     public BigDecimal getSys() {
-        return NumberUtil.div(sys * 100, total, 2);
+        return BigDecimal.valueOf(NumberUtil.div(sys * 100, total, 2));
     }
 
     public BigDecimal getUsed() {
-        return NumberUtil.div(used * 100, total, 2);
+        return BigDecimal.valueOf(NumberUtil.div(used * 100, total, 2));
     }
 
     public BigDecimal getWait() {
-        return NumberUtil.div(wait * 100, total, 2);
+        return BigDecimal.valueOf(NumberUtil.div(wait * 100, total, 2));
     }
 
     public BigDecimal getFree() {
